x411: pmufw: fix "Invalid IO Standard" warning during boot
Device-tree of zcu111 board specifies io-standard for some
PS MIO lines. Because of a bug in PMU FW kernel gives warning
message for each of those lines. The issue is described here
https://support.xilinx.com/s/article/76701?language=en_US
and there is patch to fix this bug on that page.

The patch was adapted to PMU FW 2019.2 and applied in
meta-titanium.